Dev francisco nunez

+++ src/main/java/com/msmeli/dto/request/StockBySupplierRequestDTO.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class StockBySupplierRequestDTO {
     @NotNull(message = "Ingrese un supplier id")
     @Positive(message = "Ingrese un supplier id valido")
-    private Long supplierId;
+    private Integer supplierId;
     @Valid
     @NotEmpty(message = "Ingrese al menos un item en el stock de supplier")
     private List<SupplierStockRequestDTO> content;
